proj

+++ core/io_utils.py
@@ -7,6 +7,6 @@ def read_cities(filename):
         for _ in range(num_cities):
             line = file.readline().strip().split()
             ID = int(line[0])
-            x, y = map(int, line[1:])
+            x, y = map(float, line[1:])
             cities.append(City(ID, x, y))
     return cities
